add base_resource_list.py
