+++ tests/async/fhir_server/test_async_real_fhir_server_get_patients_large.py
@@ -6,7 +6,6 @@ import pytest
 from objsize import get_deep_size
 
 from helix_fhir_client_sdk.fhir.fhir_resource_list import FhirResourceList
-from helix_fhir_client_sdk.fhir.fhir_resource_map import FhirResourceMap
 from helix_fhir_client_sdk.fhir_client import FhirClient
 from helix_fhir_client_sdk.responses.fhir_get_response import FhirGetResponse
 from helix_fhir_client_sdk.responses.fhir_merge_response import FhirMergeResponse
@@ -95,7 +94,7 @@ async def test_async_real_fhir_server_get_patients_large(
         assert response.response_headers is not None
         assert "Transfer-Encoding:chunked" in response.response_headers
         assert "Content-Encoding:gzip" in response.response_headers
-        resources: FhirResourceList | FhirResourceMap = response.get_resources()
+        resources: FhirResourceList = response.get_resources()
         assert isinstance(resources, FhirResourceList)
 
         assert len(resources) == count
